backend/python/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python>=4.8.1.78
numpy>=1.24.3
tensorflow>=2.14.0
scikit-learn>=1.3.2
colorthief>=0.2.1
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 